feat: update ingestor and shared packages
