+++ apps/ingestor/package.json
@@ -15,6 +15,7 @@
     "test:cov": "jest --coverage"
   },
   "dependencies": {
+    "@oracle-stocks/shared": "*",
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
