+++ apps/ingestor/src/providers/index.ts
@@ -1,0 +1,2 @@
+export { PriceProvider } from './provider.interface';
+export { MockProvider } from './mock.provider';
