+++ apps/ingestor/src/providers/provider.interface.ts
@@ -1,0 +1,6 @@
+import { RawPrice } from '@oracle-stocks/shared';
+
+export interface PriceProvider {
+  readonly name: string;
+  fetchPrices(symbols: string[]): Promise<RawPrice[]>;
+}
