+++ apps/ingestor/src/services/price-fetcher.service.ts
@@ -1,31 +1,25 @@
 import { Injectable, Logger } from '@nestjs/common';
-
-export interface RawPrice {
-  symbol: string;
-  price: number;
-  timestamp: number;
-  source: string;
-}
+import { RawPrice } from '@oracle-stocks/shared';
+import { PriceProvider, MockProvider } from '../providers';
 
 @Injectable()
 export class PriceFetcherService {
   private readonly logger = new Logger(PriceFetcherService.name);
   private rawPrices: RawPrice[] = [];
+  private readonly providers: PriceProvider[] = [];
+
+  constructor() {
+    this.providers.push(new MockProvider());
+  }
 
   async fetchRawPrices(): Promise<RawPrice[]> {
     const symbols = ['AAPL', 'GOOGL', 'MSFT', 'TSLA'];
-    const sources = ['AlphaVantage', 'YahooFinance', 'Finnhub'];
 
-    this.rawPrices = symbols.flatMap(symbol =>
-      sources.map(source => ({
-        symbol,
-        price: this.generateMockPrice(),
-        timestamp: Date.now(),
-        source,
-      }))
-    );
+    const pricePromises = this.providers.map(provider => provider.fetchPrices(symbols));
+    const results = await Promise.all(pricePromises);
+    this.rawPrices = results.flat();
 
-    this.logger.log(`Fetched ${this.rawPrices.length} raw prices from external APIs`);
+    this.logger.log(`Fetched ${this.rawPrices.length} raw prices from ${this.providers.length} provider(s)`);
     this.rawPrices.forEach(price => {
       this.logger.debug(
         `${price.source} - ${price.symbol}: $${price.price.toFixed(2)} at ${new Date(price.timestamp).toISOString()}`
@@ -37,9 +31,5 @@ export class PriceFetcherService {
 
   getRawPrices(): RawPrice[] {
     return this.rawPrices;
-  }
-
-  private generateMockPrice(): number {
-    return parseFloat((Math.random() * 1000 + 50).toFixed(2));
   }
 }
