+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 // Main entry point for the Shared package
 
-// Shared utilities and types will be exported here
+// Type exports
+export * from './types';
